requirements/tests.txt
pytest==3.1.3
pytest-selenium==1.11.0
pytest-variables==1.6.1
pytest-xdist==1.18.1
requests==2.18.1
pyotp==2.2.4
